apps/control-panel-backend/requirements.txt
# GT 2.0 Control Panel Backend Dependencies (Production)

# FastAPI Core
fastapi==0.121.2
uvicorn[standard]==0.38.0
pydantic[email]==2.12.4
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.44
alembic==1.16.2
asyncpg==0.30.0
psycopg2-binary==2.9.9

# Authentication & Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.10.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Two-Factor Authentication
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0

# Logging
structlog==23.2.0

# HTTP Client
httpx==0.28.1

# Message Queue
aio-pika==9.3.1

# Note: kubernetes removed - only used by resource-cluster
# Note: apscheduler removed - not currently imported/used
# Note: celery removed - not currently imported/used
